bruke sha256 i stedet for subject i nøkkelen
