+++ httpklient/src/main/kotlin/no/nav/aap/komponenter/httpklient/httpclient/tokenprovider/TexasM2MTokenProvider.kt
@@ -2,7 +2,6 @@ package no.nav.aap.komponenter.httpklient.httpclient.tokenprovider
 
 import com.github.benmanes.caffeine.cache.Cache
 import com.github.benmanes.caffeine.cache.Caffeine
-import com.github.benmanes.caffeine.cache.Expiry
 import io.micrometer.core.instrument.MeterRegistry
 import io.micrometer.core.instrument.binder.cache.CaffeineCacheMetrics
 import no.nav.aap.komponenter.config.requiredConfigForKey
@@ -11,8 +10,6 @@ import no.nav.aap.komponenter.httpklient.httpclient.RestClient
 import no.nav.aap.komponenter.httpklient.httpclient.post
 import no.nav.aap.komponenter.httpklient.httpclient.request.PostRequest
 import java.net.URI
-import java.time.Duration
-import java.time.LocalDateTime
 
 /**
  * Tokenprovider som benytter [Texas](https://doc.nais.io/auth/explanations/#texas).
@@ -60,18 +57,4 @@ internal class TexasM2MTokenProvider(
 
         return OidcToken(response.access_token)
     }
-}
-
-/** Expiry based on each token's own expiration time with a 30-second safety buffer. */
-internal fun tokenExpiry() = object : Expiry<Any, OidcToken> {
-    override fun expireAfterCreate(key: Any, value: OidcToken, currentTime: Long): Long {
-        val remaining = Duration.between(LocalDateTime.now(), value.expires())
-        return remaining.minus(Duration.ofSeconds(30)).toNanos().coerceAtLeast(0)
-    }
-
-    override fun expireAfterUpdate(key: Any, value: OidcToken, currentTime: Long, currentDuration: Long) =
-        expireAfterCreate(key, value, currentTime)
-
-    override fun expireAfterRead(key: Any, value: OidcToken, currentTime: Long, currentDuration: Long) =
-        currentDuration
 }
