+++ httpklient/src/main/kotlin/no/nav/aap/komponenter/httpklient/httpclient/tokenprovider/TexasOBOTokenProvider.kt
@@ -43,7 +43,7 @@ internal class TexasOBOTokenProvider(
         requireNotNull(scope) { "scope må være definert for token exchange med texas" }
         requireNotNull(currentToken) { "token må være tilstede for token exchange for texas" }
 
-        val key = "$scope:${currentToken.sub()}"
+        val key = "$scope:${sha256(currentToken.token())}"
         return cache.get(key) { fetchToken(scope, currentToken) }
     }
