+++ server/src/main/kotlin/no/nav/aap/komponenter/server/auth/TexasAuthenticationProvider.kt
@@ -6,18 +6,19 @@ import com.fasterxml.jackson.databind.DeserializationFeature
 import com.github.benmanes.caffeine.cache.Cache
 import com.github.benmanes.caffeine.cache.Caffeine
 import com.github.benmanes.caffeine.cache.Expiry
-import io.ktor.client.HttpClient
-import io.ktor.client.call.body
-import io.ktor.client.engine.cio.CIO
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.request.forms.submitForm
+import io.ktor.client.*
+import io.ktor.client.call.*
+import io.ktor.client.engine.cio.*
+import io.ktor.client.plugins.contentnegotiation.*
+import io.ktor.client.request.forms.*
 import io.ktor.http.*
 import io.ktor.http.auth.*
-import io.ktor.serialization.jackson.jackson
+import io.ktor.serialization.jackson.*
 import io.ktor.server.auth.*
 import io.ktor.server.auth.jwt.*
 import io.ktor.server.response.*
 import no.nav.aap.komponenter.config.requiredConfigForKey
+import no.nav.aap.komponenter.server.auth.TexasAuthenticationProvider.Companion.MAX_CACHE_TTL
 import org.slf4j.LoggerFactory
 import java.security.MessageDigest
 import java.time.Duration
